user 권한에 따라, 글 수정 제한 기능 추가

+++ bookbank/views.py
@@ -2,6 +2,7 @@ from django.shortcuts import render, redirect, get_object_or_404
 from .models import ReadingRecord
 from django.utils import timezone
 from .form import RecordForm
+from django.contrib.auth.decorators import login_required
 
 # Create your views here.
 
@@ -11,6 +12,7 @@ def record_list(request):
 
 # def recoed_request(request, username):
 
+@login_required(login_url='/account/login/')
 def new(request):
     if request.method == "POST":
         form = RecordForm(request.POST) #request.FILES) -> 이미지 넣으면 필요:
@@ -30,11 +32,18 @@ def detail(request, record_id):
     return render(request, 'detail.html', {'record':record})
 
 
+@login_required(login_url='/account/login/')
 def edit(request, record_id):
-    # user = request.user
-    form = RecordForm(instance = ReadingRecord.objects.get(pk = record_id))
-    return render(request, 'edit.html', {'form':form, 'record_id':record_id})
+    user = request.user
+    if user.is_authenticated and ReadingRecord.objects.get(pk=record_id).publisher == user.username:
+        form = RecordForm(instance = ReadingRecord.objects.get(pk = record_id))
+        return render(request, 'edit.html', {'form':form, 'record_id':record_id})
+    elif user.is_authenticated:
+        return detail(request, record_id)
+    else:
+        return redirect("login")
 
+@login_required(login_url='/account/login/')
 def update(request, record_id):
     update_record = get_object_or_404(ReadingRecord, pk = record_id)
     update_record.title = request.POST['title']
@@ -53,7 +62,12 @@ def update(request, record_id):
     return redirect('record_detail', update_record.id)
 
 def delete(request, record_id):
-    # user = request.user
-    delete_record = get_object_or_404(ReadingRecord, pk = record_id)
-    delete_record.delete()
-    return redirect('/')
+    user = request.user
+    if user.is_authenticated and ReadingRecord.objects.get(pk=record_id).publisher == user.username:
+        delete_record = get_object_or_404(ReadingRecord, pk = record_id)
+        delete_record.delete()
+        return redirect('/')
+    elif user.is_authenticated:
+        return detail(request, record_id)
+    else:
+        return redirect("login")
